Convert more tests to use real temporary files instead of fragile replacement of standard library functions
