+++ tests/test_auth.py
@@ -1,5 +1,4 @@
-from unittest.mock import patch, mock_open
-from contextlib import contextmanager
+from unittest.mock import patch
 
 import pytest
 
@@ -8,53 +7,7 @@ from hop import configure
 import os
 import stat
 
-
-@contextmanager
-def temp_environ(**vars):
-    """
-    A simple context manager for temporarily setting environment variables
-
-    Kwargs:
-        variables to be set and their values
-
-    Returns:
-        None
-    """
-    from os import environ
-    original = dict(environ)
-    os.environ.update(vars)
-    try:
-        yield  # no value needed
-    finally:
-        # restore original data
-        os.environ.clear()
-        os.environ.update(original)
-
-
-@contextmanager
-def temp_config(data, perms=stat.S_IRUSR | stat.S_IWUSR):
-    """
-    A context manager which creates a temporary config file with specified data and permissions
-
-    Args:
-        data: the data to be written to the file
-        perms: the permissions which should be set on the file.
-            The default value is to use the standard, safe permissions
-
-    Returns:
-        None
-    """
-    config_path = configure.get_config_path()
-    os.makedirs(os.path.dirname(config_path), exist_ok=True)
-    config_file = open(config_path, mode='w')
-    os.chmod(config_path, perms)
-    config_file.write(data)
-    config_file.close()
-    try:
-        yield  # no value needed
-    finally:
-        # remove file
-        os.remove(config_path)
+from conftest import temp_environ, temp_config
 
 
 def test_load_auth(auth_config, tmpdir):
@@ -95,55 +48,49 @@ def test_load_auth_malformed(tmpdir):
             auth.load_auth()
 
 
-def test_load_auth_options(auth_config):
-    # SSL should be used by default
-    # The default mechanism should be SCRAM_SHA_512
-    with patch("builtins.open", mock_open(read_data=auth_config)) as mock_file, \
-            patch("os.path.exists") as mock_exists, \
-            patch("hop.auth.Auth") as auth_mock:
-        auth.load_auth()
-        assert auth_mock.called_with(ssl=True)
-        from adc.auth import SASLMethod
-        assert auth_mock.called_with(mechanism=SASLMethod.SCRAM_SHA_512)
+def test_load_auth_options(auth_config, tmpdir):
+    with temp_environ(XDG_CONFIG_HOME=str(tmpdir)):
+        # SSL should be used by default
+        # The default mechanism should be SCRAM_SHA_512
+        with temp_config(auth_config), patch("hop.auth.Auth") as auth_mock:
+            auth.load_auth()
+            assert auth_mock.called_with(ssl=True)
+            from adc.auth import SASLMethod
+            assert auth_mock.called_with(mechanism=SASLMethod.SCRAM_SHA_512)
 
-    # But it should be possible to disable SSL
-    use_plaintext = """
+        # But it should be possible to disable SSL
+        use_plaintext = """
+                           [auth]
+                           username = "username"
+                           password = "password"
+                           protocol = "SASL_PLAINTEXT"
+                           """
+        with temp_config(use_plaintext), patch("hop.auth.Auth") as auth_mock:
+            auth.load_auth()
+            assert auth_mock.called_with(ssl=False)
+
+        # An SSL CA data path should be honored
+        with_ca_data = """
                        [auth]
                        username = "username"
                        password = "password"
-                       protocol = "SASL_PLAINTEXT"
+                       ssl_ca_location = "/foo/bar/baz"
                        """
-    with patch("builtins.open", mock_open(read_data=use_plaintext)) as mock_file, \
-            patch("os.path.exists") as mock_exists, \
-            patch("hop.auth.Auth") as auth_mock:
-        auth.load_auth()
-        assert auth_mock.called_with(ssl=False)
+        with temp_config(with_ca_data), patch("hop.auth.Auth") as auth_mock:
+            auth.load_auth()
+            assert auth_mock.called_with(ssl_ca_location="/foo/bar/baz")
 
-    # An SSL CA data path should be honored
-    with_ca_data = """
-                   [auth]
-                   username = "username"
-                   password = "password"
-                   ssl_ca_location = "/foo/bar/baz"
-                   """
-    with patch("builtins.open", mock_open(read_data=with_ca_data)) as mock_file, \
-            patch("os.path.exists") as mock_exists, \
-            patch("hop.auth.Auth") as auth_mock:
-        auth.load_auth()
-        assert auth_mock.called_with(ssl_ca_location="/foo/bar/baz")
-
-    # Alternate mechanisms should be honored
-    plain_mechanism = """
-                      [auth]
-                      username = "username"
-                      password = "password"
-                      mechanism = "PLAIN"
-                      """
-    with patch("builtins.open", mock_open(read_data=plain_mechanism)) as mock_file, \
-            patch("os.path.exists") as mock_exists, \
-            patch("hop.auth.Auth") as auth_mock:
-        auth.load_auth()
-        assert auth_mock.called_with(mechanism=SASLMethod.PLAIN)
+        # Alternate mechanisms should be honored
+        plain_mechanism = """
+                          [auth]
+                          username = "username"
+                          password = "password"
+                          mechanism = "PLAIN"
+                          """
+        with temp_config(plain_mechanism), \
+                patch("hop.auth.Auth") as auth_mock:
+            auth.load_auth()
+            assert auth_mock.called_with(mechanism=SASLMethod.PLAIN)
 
 
 def test_setup_auth(script_runner, tmpdir):
